backend/requirements.txt
argon2-cffi==21.3.0
asgiref==3.6.0
celery==5.2.7
coverage==6.4.1
Django==4.2.7
django-allauth==0.58.2
django-anymail[sendgrid]==8.6
django-cors-headers==3.14.0
django-csp==3.7
django-environ==0.8.1
django-extensions==3.1.3
django-filter==22.1
django-permissions-policy==4.13.0
django-redis==5.2.0
django-silk==5.0.3
django-stubs==1.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
djangorestframework-stubs==1.4.0
drf-dynamic-fields==0.4.0
drf-spectacular==0.26.5
factory-boy==3.2.1
faker==8.12.1
flake8==3.9.2
gunicorn==20.1.0
ipdb==0.13.9
mypy==1.1.1
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pylint==2.14.0
pylint-django==2.5.3
pytest==6.2.5
pytest-django==4.4.0
pytest-sugar==0.9.4
pytz==2022.6
qrcode[pil]==7.3.1
requests==2.28.2
sendgrid==6.10.0
sqlparse==0.4.2
unidecode==1.3.6
uvicorn[standard]==0.17.6
weasyprint==57.1
